Coerced text drawing to pixel boundaries - basically just cast to int when creating a Vector2 to pass to SpriteBatch.DrawString().  This chops off the fraction and moves the text to the top left of the pixel resulting in much clearer text rendering.

+++ XPF/RedBadger.Xpf.Adapters.Xna/Graphics/SpriteBatchAdapter.cs
@@ -81,7 +81,7 @@
                 this.spriteBatch.DrawString(
                     spriteFontAdapter.Value, 
                     text ?? string.Empty, 
-                    new Vector2((float)position.X, (float)position.Y), 
+                    new Vector2((int)position.X, (int)position.Y),
                     new Microsoft.Xna.Framework.Color(color.R, color.G, color.B, color.A));
             }
         }
